CHET-141: Create bean sepearate configuration classes. Use component scans instead of referring to the configuration class dierctly
